+++ tests/helpers/destroy-app.js
@@ -1,5 +1,6 @@
 import Ember from 'ember';
 
 export default function destroyApp(application) {
+  Ember.run.cancel(window._breakTimerLoopsId);
   Ember.run(application, 'destroy');
 }